ändrade id i alla DAO klasser till short

+++ src/main/java/Business/Category.java
@@ -17,7 +17,7 @@ public class Category {
     @Column(name = "last_update")
     private Timestamp lastUpdate;
 
-    public Category(int categoryId, String name, Timestamp lastUpdate) {
+    public Category(byte categoryId, String name, Timestamp lastUpdate) {
         this.categoryId = categoryId;
         this.name = name;
         this.lastUpdate = lastUpdate;
